requirements.txt
Pillow
pytesseract
natsort
